refactor: removed 9 unused components and updated file structure
