+++ src/components/ProductListing/AccessoryList/FilterSortPanel__OLD/Accordion.jsx
@@ -1,9 +1,9 @@
 import { useState, useRef, useEffect } from "react";
 import PropTypes from "prop-types";
 import styled from "@emotion/styled";
-import chevron from "../../../images/new/chevron-down-solid.svg";
-import { useWindowSize } from "../../../contexts/WindowSize.ctx";
-import { colors, fonts } from "../../../styles/theme.styled";
+import chevron from "../../../../images/new/chevron-down-solid.svg";
+import { useWindowSize } from "../../../../contexts/WindowSize.ctx";
+import { colors, fonts } from "../../../../styles/theme.styled";
 
 const AccordionWrapper = styled.div`
   margin-bottom: 10px;
