+++ src/components/ProductListing/AccessoryList/FilterSortPanel__OLD/FilterByCard.jsx
@@ -2,7 +2,7 @@ import { Children } from "react";
 import PropTypes from "prop-types";
 
 import styled from "@emotion/styled";
-import { colors, fonts } from "../../../styles/theme.styled";
+import { colors, fonts } from "../../../../styles/theme.styled";
 
 const FilterByCardRoot = styled.div`
   background-color: white;
