--- src/components/ProductListing/AllMattressList/index.js
@@ -1,1 +1,0 @@
-export { default } from "./AllMattressList";
