+++ src/components/ProductListing/MattressList/generateInitialState.js
