+++ src/components/ProductListing/MattressList/helperFunctions.js
