--- src/components/ProductListing/ProductList/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ProductList";
