+++ src/pages/brands/list.js
@@ -2,8 +2,8 @@ import { graphql } from "gatsby";
 import { HelmetDatoCms } from "gatsby-source-datocms";
 import propTypes from "prop-types";
 import Layout from "../../components/Layout";
-import QueryStringFilterState from "../../components/ProductListing/Mattresses/QueryStringFilterState";
-import ProductList from "../../components/ProductListing/ProductList";
+import MattressListInit from "../../components/ProductListing/MattressList/InitializeLists/MattressListInit";
+import ProductList from "../../components/ProductListing/MattressList";
 import getMattresses from "../../components/ProductListing/QueryHooks/Mattresses/Content/getAll.query";
 import getHeaders from "../../components/ProductListing/QueryHooks/Mattresses/Content/getHeaders.query";
 
@@ -44,4 +44,4 @@ List.propTypes = {
   location: propTypes.instanceOf(Object).isRequired,
   initialFilterState: propTypes.instanceOf(Object).isRequired,
 };
-export default QueryStringFilterState(List);
+export default MattressListInit(List);
